chore(tel-ui): Move MQTT types to ui package
