+++ telemetry/packages/types/src/index.ts
@@ -8,9 +8,3 @@ export type {
   OpenMctObjectTypes,
   OpenMctObjectType,
 } from './openmct/openmct-object-types.types';
-export type {
-  QoS,
-  MqttPublish,
-  MqttSubscribe,
-  MqttUnsubscribe,
-} from './mqtt/mqtt.types';
